New dir: internal/services/autorization new file autorization/autorization.go - midlweare functionб taskdecode.go taskencode.go -> rename package
